Cast negative numbers before assigning them to unsigned variables to prevent compiler complaints.
git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@15565 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXportLib/src/os/OsFileBase.cpp
@@ -402,7 +402,7 @@ OsStatus OsFileBase::getPosition(size_t &pos)
 {
    OsLock lock(fileMutex);
 
-   pos = UTL_NOT_FOUND;
+   pos = (size_t) UTL_NOT_FOUND;
 
 #ifdef DEBUG_FS
    pthread_t nTaskId = 0;
@@ -410,6 +410,7 @@ OsStatus OsFileBase::getPosition(size_t &pos)
    if (pTask) pTask->id(nTaskId);
    OsSysLog::add(FAC_KERNEL, PRI_DEBUG, "OsFileBase::getPosition ENTER threadid=%ld\n", nTaskId);
 #endif
+
     OsStatus stat = OS_INVALID;
 
     if (mOsFileHandle)
@@ -424,6 +425,7 @@ OsStatus OsFileBase::getPosition(size_t &pos)
 #ifdef DEBUG_FS
    OsSysLog::add(FAC_KERNEL, PRI_DEBUG, "OsFileBase::getPosition EXIT threadid=%ld\n", nTaskId);
 #endif
+
     return stat;
 }
